Enable Renovate patch updates for react-router
This change tells Renovate to upgrade react-router even if the patch
version is the only thing that changes. For example, react-router is
currently at version 3.2.0. If version 3.2.1 is released, Renovate will
upgrade to that version.

Renovate upgrades packages when the minor version changes (e.g., 3.3.0)
by default.

I think it's a good idea to turn this on for react-router because the
3.x.x version of react-router is rarely updated, so we want to get any
updates that become available. We decided not to upgrade to the 4.x.x
version, which gets more frequent updates, because the work to do that
upgrade is non-trivial and the react-router team said that they they
will support 3.x.x indefinitely.

More information:
https://renovateapp.com/docs/all-other/faq#separate-patch-releases-from-minor-releases